moved new adapters to gui/networkadapters and deleted the obsolete corresponding jobs from newwizard/jobs
