+++ src/gui/networkadapters/resolveurladapter.cpp
