+++ src/gui/networkadapters/resolveurladapter.h
